Fix parachain bridge init bridge
