+++ frame/assistants/client-pangolin/src/fastapi/bridge_s2s/bridge_pangoro.rs
@@ -107,7 +107,7 @@ impl S2SClientRelay for PangolinClient {
             .await?;
         let events = track.wait_for_finalized_success().await.map_err(|e| {
             S2SClientError::RPC(format!(
-                "send transactioni failed {}: {:?}",
+                "send transaction failed {}: {:?}",
                 <Self as S2SClientBase>::CHAIN,
                 e
             ))
@@ -133,7 +133,7 @@ impl S2SClientRelay for PangolinClient {
             .await?;
         let events = track.wait_for_finalized_success().await.map_err(|e| {
             S2SClientError::RPC(format!(
-                "send transactioni failed {}: {:?}",
+                "send transaction failed {}: {:?}",
                 <Self as S2SClientBase>::CHAIN,
                 e
             ))
@@ -213,7 +213,7 @@ impl S2SClientRelay for PangolinClient {
             .await?;
         let events = track.wait_for_finalized_success().await.map_err(|e| {
             S2SClientError::RPC(format!(
-                "send transactioni failed {}: {:?}",
+                "send transaction failed {}: {:?}",
                 <Self as S2SClientBase>::CHAIN,
                 e
             ))
@@ -239,7 +239,7 @@ impl S2SClientRelay for PangolinClient {
             .await?;
         let events = track.wait_for_finalized_success().await.map_err(|e| {
             S2SClientError::RPC(format!(
-                "send transactioni failed {}: {:?}",
+                "send transaction failed {}: {:?}",
                 <Self as S2SClientBase>::CHAIN,
                 e
             ))
